Add cli flag for custom ipc path
