+++ Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+clap = "2.33.3"
 reqwest = { version = "0.11.0", features = ["blocking"] }
 serde_json = "1.0.59"
 threadpool = "1.8.1"
-clap = "2.33.3"
